Systemic risk measures
This code estimates the MES, SRISK and CoVaR using GARCH-DCC model.
These popular market-based systemic risk measures capture the
contribution of a given financial institution to the risk of the system.
The first two measures reproduce quantities introduced by Brownlees and
Engle (2012) and available on the V-Lab website for a large range of
financial institutions. The last measure is also estimated with a
quantile regression as proposed by Adrian and Brunnermeier (2011).